+++ src/stactools/cop_dem/constants.py
@@ -14,7 +14,7 @@ COP_DEM_COLLECTION_END: Optional[datetime] = str_to_datetime(
     "2021-04-22T00:00:00Z")
 COP_DEM_TEMPORAL_EXTENT = [COP_DEM_COLLECTION_START,
                            COP_DEM_COLLECTION_END]  # TODO: find the dates
-COP_DEM_PLATFORM = "TanDEM-X"
+COP_DEM_PLATFORM = "tandem-x"
 COP_DEM_EPSG = 4326
 COP_DEM_KEYWORDS = ['DEM', 'COPERNICUS', 'DSM', 'Elevation']
 COP_DEM_PROVIDERS = [